Move the tokenizer back to its original location
